Fix warnings

+++ include/Game/RELMOD.h
@@ -1,7 +1,7 @@
 #ifndef _RELMOD_H_
 #define _RELMOD_H_
 
-void RELMOD_RelocModulePointers(int baseaddr, int offset, int *relocs);
-void RELMOD_InitModulePointers(int baseaddr, int *relocs);
+void RELMOD_RelocModulePointers(intptr_t baseaddr, int offset, int *relocs);
+void RELMOD_InitModulePointers(intptr_t baseaddr, int *relocs);
 
 #endif
